requirements.txt
streamlit
torch
torchvision
torchaudio
google-cloud-vision